Form 194/style guide brands (#337)
* feat: separate Wrapper and Stack into independent classes

* refactor: divide code into separate showcases

* feat: create StyleGuideWrapper where will be showcases held and also will be able to change brand style

* feat: add changing of brand color scheme in styleguide with buttons

* style: add border on top, add paddings and some more details
